refactor: move packages from header to pkg/header
