+++ nodebuilder/state/core.go
@@ -2,8 +2,8 @@ package state
 
 import (
 	apptypes "github.com/celestiaorg/celestia-app/x/payment/types"
-	"github.com/celestiaorg/celestia-node/header/sync"
 	"github.com/celestiaorg/celestia-node/nodebuilder/core"
+	"github.com/celestiaorg/celestia-node/pkg/header/sync"
 	"github.com/celestiaorg/celestia-node/state"
 )
