+++ pkg/header/p2p/exchange_test.go
@@ -22,8 +22,8 @@ import (
 	swarm "github.com/libp2p/go-libp2p/p2p/net/swarm/testing"
 
 	"github.com/celestiaorg/celestia-node/header"
-	headerMock "github.com/celestiaorg/celestia-node/header/mocks"
-	p2p_pb "github.com/celestiaorg/celestia-node/header/p2p/pb"
+	headerMock "github.com/celestiaorg/celestia-node/pkg/header/mocks"
+	p2p_pb "github.com/celestiaorg/celestia-node/pkg/header/p2p/pb"
 )
 
 var privateProtocolID = protocolID("private")
