+++ pkg/header/p2p/helpers.go
@@ -13,7 +13,7 @@ import (
 	"github.com/celestiaorg/go-libp2p-messenger/serde"
 
 	"github.com/celestiaorg/celestia-node/header"
-	p2p_pb "github.com/celestiaorg/celestia-node/header/p2p/pb"
+	p2p_pb "github.com/celestiaorg/celestia-node/pkg/header/p2p/pb"
 )
 
 func protocolID(protocolSuffix string) protocol.ID {
