+++ pkg/header/p2p/server.go
@@ -16,8 +16,8 @@ import (
 	"github.com/celestiaorg/go-libp2p-messenger/serde"
 
 	"github.com/celestiaorg/celestia-node/header"
-	p2p_pb "github.com/celestiaorg/celestia-node/header/p2p/pb"
 	headerpkg "github.com/celestiaorg/celestia-node/pkg/header"
+	p2p_pb "github.com/celestiaorg/celestia-node/pkg/header/p2p/pb"
 )
 
 var (
