+++ pkg/header/store/batch.go
