+++ pkg/header/store/height_indexer.go
