+++ pkg/header/sync/ranges.go
